# docs/swagger.json
{"openapi":"3.0.0","paths":{"/v1/app/health":{"get":{"operationId":"AppController_health","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["app"]}},"/v1/auth/token":{"get":{"operationId":"AuthController_createToken","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponseDto"}}}}},"tags":["auth"]}},"/v1/private":{"get":{"operationId":"PrivateController_getPrivates","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PrivateResponseDto"}}}}}},"tags":["private"],"security":[{"bearer":[]}]},"post":{"operationId":"PrivateController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateInputDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateResponseDto"}}}}},"tags":["private"],"security":[{"bearer":[]}]}},"/v1/private/{id}":{"get":{"operationId":"PrivateController_getPrivate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateResponseDto"}}}}},"tags":["private"],"security":[{"bearer":[]}]},"patch":{"operationId":"PrivateController_updateById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateInputDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateResponseDto"}}}}},"tags":["private"],"security":[{"bearer":[]}]},"delete":{"operationId":"PrivateController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["private"],"security":[{"bearer":[]}]}},"/v1/public":{"get":{"operationId":"PublicController_getPublics","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PublicResponseDto"}}}}}},"tags":["public"]},"post":{"operationId":"PublicController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicInputDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicResponseDto"}}}}},"tags":["public"]}},"/v1/public/{id}":{"get":{"operationId":"PublicController_getPublic","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicResponseDto"}}}}},"tags":["public"]},"patch":{"operationId":"PublicController_updateById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicInputDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PublicResponseDto"}}}}},"tags":["public"]},"delete":{"operationId":"PublicController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["public"]}}},"info":{"title":"Outvio Nest App","description":"Nest Prisma Mongo Rest API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"TokenResponseDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"PrivateResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt"]},"PrivateInputDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"PublicResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","createdAt","updatedAt"]},"PublicInputDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]}}}}