# src/modules/private/dto/private.input.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

export class PrivateInputDto {
  @ApiProperty()
  @IsString()
  @IsNotEmpty()
  name!: string;
}